% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{calculate_stem_area}
\alias{calculate_stem_area}
\title{Calculate mid-point stem area from tree size class}
\usage{
calculate_stem_area(sizes, counts = NULL, units = c("m2", "cm2"))
}
\arguments{
\item{sizes}{A factor or character vector of tree size classes as per the
'sizeclass' column in the \code{\link{SIZES}} table (e.g. 'from20to30cm').}

\item{counts}{An option vector of tree counts. If provided it must be either
the same length as 'sizes' or a single value. If not provided, counts of
1 are assumed.}

\item{units}{Area units: one of 'm2' (default) or 'cm2'.}
}
\value{
A vector of stem area values.
}
\description{
This function uses representative mid-point diameters for tree size classes
(as defined in the \code{\link{SIZES}} table) to calculate stem areas.
}
