% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_functions.R
\name{jags_cell_means_gamma}
\alias{jags_cell_means_gamma}
\title{Fit a cell-means model with JAGS for a Gamma distributed response variable}
\usage{
jags_cell_means_gamma(.data, yvar, cellvar = "cell", re.idvar = "re.id",
  n.chains = 2, n.update = 1000, n.sims = 5000,
  variable.names = c("mu.cell", "re.sigma", "shape"),
  gelman.threshold = 1.1, gelman.action = c("stop", "warning"))
}
\arguments{
\item{.data}{Model data with columns for response, cell identifier and
random effect identifier.}

\item{yvar}{Name or column index for response variable.}

\item{cellvar}{Name or column index for integer cell identifier. Values are
assumed to be consecutive integers from 1.}

\item{re.idvar}{Name or column index for integer random effect identifier.
Values are assumed to be consecutive integers from 1.}

\item{n.chains}{Number of Markov chain Monte Carlo chains (default 2).}

\item{n.update}{Number of updating iterations to run (default 1000).}

\item{n.sims}{Number of sampling iterations to run (default 5000).}

\item{variable.names}{Names of model variables to monitor. Defaults to
c("mu.cell", "re.sigma", "shape").}

\item{gelman.threshold}{Value of the Gelman-Rubin convergence diagnostic to
use when assessing convergence of sampled variables (default 1.1).}

\item{gelman.action}{Action if one or more variables do not pass the Gelman-Rubin
convergence check. Options are 'stop' (default) or 'warning'.}
}
\value{
An MCMC object of sample values as a single chain.
}
\description{
Fit a cell-means model with JAGS for a Gamma distributed response variable
}
