% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAGESIZE}
\alias{PAGESIZE}
\title{Page size information for use with the \code{\link{save_plot}} function.}
\format{An object of class \code{list} of length 2.}
\usage{
PAGESIZE
}
\description{
A named list with elements 'A3' and 'A4'. Each element is a list with
elements 'width', 'height' and 'units'.
}
\keyword{datasets}
