% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{plain_percent}
\alias{plain_percent}
\title{Format axis tick labels as percentages}
\usage{
plain_percent(x)
}
\arguments{
\item{x}{Numeric vector of values to format.}
}
\value{
Labels for `x` values multiplied by 100.
}
\description{
This function is a wrapper around `scales::percent` to
give tick labels without the '%' symbol.
}
\examples{
plain_percent(runif(10))

}
