% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot object to file.}
\usage{
save_plot(ggobj, path, pagesize, orientation = c("portrait", "landscape"))
}
\arguments{
\item{ggobj}{The ggplot object to save.}

\item{path}{The path and name of the destination file. This should include the extension
(e.g. 'foo.pdf') as this is used to determine the file format.}

\item{pagesize}{A list of page size information with elements 'width', 'height' and
'units'. Use the \code{\link{PAGESIZE}} object for A3 and A4 information.}

\item{orientation}{One of 'portrait' or 'landscape' (may be abbreviated).}
}
\description{
Save a ggplot object to file.
}
\examples{
\dontrun{
save_plot(last_plot(), "treedata.pdf", PAGESIZE$A4, orientation = "land")
}

}
