% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SITES}
\alias{SITES}
\title{RRG site information}
\format{An object of class \code{data.frame} with 66 rows and 5 columns.}
\usage{
SITES
}
\description{
A data frame of meta-data for sites and plots consisting of:
\describe{
\item{site}{Site identifier (int)}
\item{plot}{Plot identifier (int)}
\item{site.quality}{Mapped site quality class (factor: SQ1; SQ2)}
\item{year.thinned}{Year of thinning for treatment plots (factor: 2016; 2017)}
\item{treat}{Plot treatment (factor: control; moderate; heavy)}
}
}
\keyword{datasets}
