% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIZES}
\alias{SIZES}
\title{RRG tree diameter size classes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 4 columns.}
\usage{
SIZES
}
\description{
A data frame of size class definitions and identifiers consisting of:
\describe{
\item{sizeclass}{Column names in tree count data sets
  (factor with 11 levels)}
\item{index}{Integer size class indices}
\item{label}{Alternative short labels for use in graphs etc.
  (factor with 11 levels)}
}
}
\keyword{datasets}
